Update 6502.c - cast 64bit to ulong

+++ src/sim65/6502.c
@@ -4809,7 +4809,7 @@ unsigned long GetCycles (void)
 /* Return the total number of cycles executed */
 {
     /* Return the total number of cycles */
-    return Peripherals.Counter.ClockCycles;
+    return (unsigned long)Peripherals.Counter.ClockCycles;
 }
